# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/comments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "对作品发表评论或回复",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "发表评论",
                "parameters": [
                    {
                        "description": "评论请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "评论成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误、作品不存在、父评论不存在或不能回复二级及以下的评论",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/user/{user_id}": {
            "get": {
                "description": "获取指定用户的所有评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "获取用户评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的用户ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/work/{work_id}": {
            "get": {
                "description": "获取指定作品的所有评论（带分页）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "获取作品评论列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "work_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的作品ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/work/{work_id}/root": {
            "get": {
                "description": "获取指定作品的一级评论（不带分页，返回树形结构）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "获取作品根评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "work_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的作品ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{id}": {
            "get": {
                "description": "根据ID获取评论详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "获取评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的评论ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "评论不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新评论内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "更新评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "评论内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "评论不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的评论ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "评论不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{id}/like": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新评论的点赞数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "更新评论点赞数",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "增量",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "点赞更新成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "评论不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新评论审核状态（仅管理员可操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "更新评论状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态: 0删除 1正常 2审核中",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "状态更新成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或无效的评论状态",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "评论不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backend_internal_models.CreateCommentRequest": {
            "type": "object",
            "required": [
                "content",
                "work_id"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 1000
                },
                "media_url": {
                    "type": "string"
                },
                "parent_id": {
                    "description": "回复某条评论",
                    "type": "integer"
                },
                "root_id": {
                    "description": "所属楼层根 ID",
                    "type": "integer"
                },
                "work_id": {
                    "type": "integer"
                }
            }
        },
...
        "backend_pkg_response.ErrorField": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "backend_pkg_response.PageInfo": {
            "type": "object",
            "properties": {
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "backend_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/backend_pkg_response.ErrorField"
                },
                "page": {
                    "$ref": "#/definitions/backend_pkg_response.PageInfo"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
    }
}